requirements.txt
# Core dependencies
pyaudio>=0.2.13
webrtcvad>=2.0.10
vosk==0.3.44
numpy>=1.24.0,<2.3.0  # Pinned: numba (used by mlx-whisper) requires numpy <= 2.3
phonetics>=1.0.5
psutil>=5.9.0  # For memory monitoring
pynput>=1.7.6  # For hotkey functionality

# ASR (Speech-to-Text) dependencies
mlx-whisper>=0.4.0
parakeet-mlx>=0.1.0  # For Parakeet transcription models
# Note: parakeet-mlx requires Python 3.10+ and may not be available on all systems
# If installation fails, the app will fall back to Whisper models
huggingface-hub>=0.23.0
mlx-audio>=0.2.4  # For Voxtral MLX ASR models
librosa>=0.10.0  # For MedASR audio loading

# MedASR dependencies (for google/medasr model)
transformers>=4.40.0  # For MedASR pipeline
torch>=2.0.0  # Required by transformers for MedASR

# Audio conversion
pydub>=0.25.1

# Cloud ASR backends
httpx>=0.27.0
google-cloud-speech>=2.24.0  # optional, for Google Chirp 2

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xvfb>=3.0.0
mock>=5.0.0

# Code quality dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
bandit>=1.7.0
safety>=2.0.0

# Development dependencies
# pylint>=2.15.0
